// src/app/page/master/metode-pembayaran/metode-pembayaran.interface.ts
import {EventEmitter} from '@angular/core';

export interface MetodePembayaran {
	kode?: any;
	id?: any;
	mataUang?: any;
	kdProfile?: string;
	namaModulAplikasi?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}